+++ src/background.ts
@@ -555,14 +555,18 @@ and make sure the answer is not longer than the user's question. You always answ
 // Summarizer API
 async function initializeSummarizer() {
   if (typeof Summarizer === 'undefined') {
-    notifyPopup('summarizer', false, 'Summarizer not available');
+    notifyPopup(
+      'summarizer',
+      false,
+      'Summarizer not available in this version of Chrome'
+    );
     console.warn('LanguageModel API not available in this context.');
     return;
   }
 
   const availability = await Summarizer.availability();
   if (availability !== 'available') {
-    notifyPopup('summarizer', false, 'Summarizer not ready (downloading...)');
+    notifyPopup('summarizer', false, 'Summarizer not available');
   }
   try {
     // Proceed to request batch or streaming summarization
@@ -603,7 +607,11 @@ async function initializeSummarizer() {
 
 async function initializeTranslator() {
   if (typeof Translator === 'undefined') {
-    notifyPopup('translator', false, 'Translator not available');
+    notifyPopup(
+      'translator',
+      false,
+      'Translator not available in this version of Chrome'
+    );
     console.warn('Translator API not available in this context.');
     return;
   }
@@ -614,7 +622,7 @@ async function initializeTranslator() {
   });
 
   if (availability !== 'available') {
-    notifyPopup('translator', false, 'Translator not ready (downloading...)');
+    notifyPopup('translator', false, 'Translator not available');
   }
   try {
     session_t_jp_en = await Translator.create({
@@ -658,7 +666,11 @@ async function initializeTranslator() {
 
 async function initializeLanguageModel() {
   if (typeof LanguageModel === 'undefined') {
-    notifyPopup('language-model', false, 'LanguageModel not available');
+    notifyPopup(
+      'language-model',
+      false,
+      'LanguageModel not available in this version of Chrome'
+    );
     console.warn('LanguageModel API not available in this context.');
     return;
   }
@@ -666,11 +678,7 @@ async function initializeLanguageModel() {
   const availability = await LanguageModel.availability();
 
   if (availability !== 'available') {
-    notifyPopup(
-      'language-model',
-      false,
-      'LanguageModel not ready, downlaoding...'
-    );
+    notifyPopup('language-model', false, 'LanguageModel not available');
   }
 
   try {
@@ -746,7 +754,7 @@ function safeParseModelJson(raw: string) {
     return [];
   }
 }
-type Feature = 'summarizer' | 'translator' | 'language-model';
+
 interface ChatMessage {
   user: 'me' | 'ai';
   text: string;