+++ src/utils/types/global.d.ts
@@ -1,0 +1,5 @@
+declare global {
+  type Feature = 'summarizer' | 'translator' | 'language-model';
+}
+
+export {};