+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # 0.1.0 - 0.1.12
+
+## 0.13
+1. Add utils/types to collect types
+2. Add moveFiles.js to move important files for uploading to Chrome Web Store
+3. UI improvement on prev/next button
+4. Add TODO on percentage UI feature
+
 ## 0.12
 1. Every tab changes will get featureStatus from extension storage
 2. Add CHANGELOG.md